Mejora al editar un rol

+++ usuarios/bd_update/actualizar-usuarios.php
@@ -25,23 +25,24 @@ while ($fila = $consultaRolesUsuario->fetch_assoc()) {
     $rolesExistentes[] = $fila['upr_id_rol'];
 }
 
-$rolesAEliminar = array_diff($rolesExistentes, $rolesSeleccionados);
-$rolesAInsertar = array_diff($rolesSeleccionados, $rolesExistentes);
+if (!empty($_POST['tipoU'])) {
+    $rolesAEliminar = array_diff($rolesExistentes, $rolesSeleccionados);
+    $rolesAInsertar = array_diff($rolesSeleccionados, $rolesExistentes);
 
-if (!empty($rolesAEliminar)) {
-    $rolesAEliminarStr = implode("','", $rolesAEliminar);
-    $conexionBdAdmin->query("DELETE FROM usuarios_roles WHERE upr_id_usuario = '$idUsuario' AND upr_id_rol IN ('$rolesAEliminarStr')");
-}
+    if (!empty($rolesAEliminar)) {
+        $rolesAEliminarStr = implode("','", $rolesAEliminar);
+        $conexionBdAdmin->query("DELETE FROM usuarios_roles WHERE upr_id_usuario = '$idUsuario' AND upr_id_rol IN ('$rolesAEliminarStr')");
+    }
 
-$upr_fec = date("Y-m-d H:i:s");
-$upr_responsable = $_SESSION["id"];
-$id_empresa = $_SESSION["dataAdicional"]["id_empresa"];
-if (!empty($rolesAInsertar)) {
-    foreach ($rolesAInsertar as $rol) {
-        $conexionBdAdmin->query("INSERT INTO usuarios_roles (upr_id_usuario, upr_id_rol, upr_fec, upr_responsable,upr_id_empresa) VALUES ('$idUsuario', '$rol', '$upr_fec', '$upr_responsable', '$id_empresa')");
+    $upr_fec = date("Y-m-d H:i:s");
+    $upr_responsable = $_SESSION["id"];
+    $id_empresa = $_SESSION["dataAdicional"]["id_empresa"];
+    if (!empty($rolesAInsertar)) {
+        foreach ($rolesAInsertar as $rol) {
+            $conexionBdAdmin->query("INSERT INTO usuarios_roles (upr_id_usuario, upr_id_rol, upr_fec, upr_responsable,upr_id_empresa) VALUES ('$idUsuario', '$rol', '$upr_fec', '$upr_responsable', '$id_empresa')");
+        }
     }
 }
-
 
 include(RUTA_PROYECTO."/usuarios/includes/guardar-historial-acciones.php");
